// src/app/models/users.model.ts
export class UsersModel {
  firstName: string;
  lastName: string;
  displayName: string;
  username: string;
  email: string;
  password: string;
  returnSecureToken: boolean;
  method: string;
  picture: string;
  idToken: string;
  needConfirm: boolean;
  country: string;
  countryCode: string;
  city: string;
  address: string;
  phone: string;
}